refactor(Lotto): Lotto의 패키지는 변경할 수 없다.
