+++ src/main/java/lotto/domain/PrizeLevel.java
@@ -1,11 +1,11 @@
 package lotto.domain;
 
 public enum PrizeLevel {
-    MATCH_3("3개 일치 (5,000원)", 3, false, 5000),
-    MATCH_4("4개 일치 (50,000원)", 4, false, 50000),
-    MATCH_5("5개 일치 (1,500,000원)", 5, false, 1500000),
-    MATCH_5_BONUS("5개 일치, 보너스 볼 일치 (30,000,000원)", 5, true, 30000000),
-    MATCH_6("6개 일치 (2,000,000,000원)", 6, false, 2000000000);
+    MATCH_3("3개 일치 (5,000원)", 3, false, 5_000),
+    MATCH_4("4개 일치 (50,000원)", 4, false, 50_000),
+    MATCH_5("5개 일치 (1,500,000원)", 5, false, 1_500_000),
+    MATCH_5_BONUS("5개 일치, 보너스 볼 일치 (30,000,000원)", 5, true, 30_000_000),
+    MATCH_6("6개 일치 (2,000,000,000원)", 6, false, 2_000_000_000);
 
     private final String displayName;
     private final int matchCount;
