+++ src/test/java/lotto/unit/ResultTest.java
@@ -6,8 +6,8 @@ import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 import java.util.Arrays;
 import java.util.List;
+import lotto.Lotto;
 import lotto.constants.OutputMessages;
-import lotto.domain.Lotto;
 import lotto.domain.Result;
 import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
